modelAPI/requirements.txt
Flask==2.3.3
requests==2.31.0
tensorflow
pillow==10.3.0
debugpy